requirements.txt
streamlit
streamlit_chat
pinecone-client
google-generativeai 
langchain-google-genai 
